sneaker-processing-api/requirements.txt
flask
numpy
opencv-contrib-python
Pillow